j'ai essayé d'optimiser un peu plus. Comme l'avait suggéré Paul, je laisse le A* renvoyé la meme position jusqu'a que je l'atteignes, ce qui permet des mouvements plus fluides et une interruption d'un mouvement en cas d'opportunité d'attaque ou d'evasion. Le probleme c'est que cette version a plus de crash de la simu. A voir les differences entre avant et maintenant

+++ Strategy/Strategy.h
@@ -39,9 +39,9 @@ public:
 
     void strategy();
 
-    bool CloseEnemy(float dist_thresh);
+    bool CloseEnemy(uint8_t dist_thresh);
 
-    bool CloseDeadEnemy(float dist_thresh);
+    bool CloseDeadEnemy(uint8_t dist_thresh);
 
     bool CloseMaxWall();
